Split active detailed status presenters
